Move (u)int_impl macros to their own files
